Improves Dockerfile
The dockerfile does not follow the best practices with respect to caching.
The requirements.txt should be copied first and the environment be set up.
This helps in scenarios when the project dependencies don't change but the
code changes. Before this commit, all the layers would have run resulting
in slower build times.